backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic-ai-slim[google]>=0.0.14
python-dotenv>=1.0.1
requests>=2.32
# fastembed usa ONNX Runtime (sem torch). Se nao instalado, o backend cai para BM25.
fastembed>=0.4.0
